extract sandbox kill POST into SWR hook
