+++ front/components/poke/PokeNavbar.tsx
@@ -67,11 +67,6 @@ function PokeNavbar({ regionUrls, showRegionPicker = false }: PokeNavbarProps) {
           <Button href="/poke/templates" variant="ghost" label="Templates" />
           <Button href="/poke/plugins" variant="ghost" label="Plugins" />
           <Button href="/poke/kill" variant="ghost" label="Kill Switches" />
-          <Button
-            href="/poke/sandbox-kill"
-            variant="ghost"
-            label="Sandbox Kill"
-          />
           <Button href="/poke/cache" variant="ghost" label="Cache" />
           <Button href="/poke/pokefy" variant="ghost" label="Pokefy URL" />
           <Button
