+++ front/components/poke/pages/KillPage.tsx
@@ -4,18 +4,25 @@ import { useSendNotification } from "@app/hooks/useNotification";
 import { clientFetch } from "@app/lib/egress/client";
 import { KILL_SWITCH_TYPES, type KillSwitchType } from "@app/lib/poke/types";
 import { usePokeKillSwitches } from "@app/poke/swr/kill";
+import {
+  usePokeSandboxKillImages,
+  useRequestSandboxKill,
+} from "@app/poke/swr/sandbox_kill";
 import { normalizeError } from "@app/types/shared/utils/error_utils";
 import {
   ActionFireIcon,
   AnthropicLogo,
   ArrowPathIcon,
   BoltIcon,
+  Button,
   CardIcon,
   CloudArrowLeftRightIcon,
   Cog6ToothIcon,
+  ExclamationCircleIcon,
   OpenaiLogo,
   SliderToggle,
   Spinner,
+  TrashIcon,
 } from "@dust-tt/sparkle";
 import type { ComponentType } from "react";
 import { useState } from "react";
@@ -70,6 +77,26 @@ const KILL_SWITCH_DEFINITIONS: Record<KillSwitchType, KillSwitchDefinition> = {
   },
 };
 
+const PANEL_HEADING_CLASSES =
+  "flex items-center gap-2.5 text-2xl font-semibold tracking-tight text-foreground dark:text-foreground-night";
+const PANEL_ICON_CLASSES =
+  "h-4 w-4 text-muted-foreground dark:text-muted-foreground-night";
+const PANEL_DESCRIPTION_CLASSES =
+  "text-sm text-muted-foreground dark:text-muted-foreground-night";
+const PANEL_SECTION_CLASSES = cn(
+  "mt-6 rounded-2xl border border-border",
+  "bg-background shadow-sm dark:border-border-night dark:bg-background-night"
+);
+
+type SandboxKillRequestKey = string;
+
+function sandboxKillKey(
+  baseImage: string,
+  version?: string
+): SandboxKillRequestKey {
+  return `${baseImage}|${version ?? ""}`;
+}
+
 export function KillPage() {
   useDocumentTitle("Poke - Kill Switches");
 
@@ -79,6 +106,11 @@ export function KillPage() {
     useState<KillSwitchType | null>(null);
   const sendNotification = useSendNotification();
   const enabledKillSwitches = new Set(killSwitches);
+
+  const { images, isImagesLoading } = usePokeSandboxKillImages();
+  const requestSandboxKill = useRequestSandboxKill();
+  const [submittingSandboxKill, setSubmittingSandboxKill] =
+    useState<SandboxKillRequestKey | null>(null);
 
   async function updateKillSwitch(
     killSwitch: KillSwitchType,
@@ -140,82 +172,185 @@ export function KillPage() {
     }
   }
 
+  async function submitSandboxKillRequest(
+    baseImage: string,
+    version: string | undefined,
+    confirmMessage: string
+  ): Promise<void> {
+    if (submittingSandboxKill) {
+      return;
+    }
+    if (!window.confirm(confirmMessage)) {
+      return;
+    }
+
+    const key = sandboxKillKey(baseImage, version);
+    setSubmittingSandboxKill(key);
+    try {
+      await requestSandboxKill({ baseImage, version });
+    } finally {
+      setSubmittingSandboxKill(null);
+    }
+  }
+
   return (
-    <main className="mx-auto max-w-4xl px-4 py-8 sm:px-6 lg:px-8">
-      <header className="space-y-2">
-        <h1 className="flex items-center gap-2.5 text-2xl font-semibold tracking-tight text-foreground dark:text-foreground-night">
-          <BoltIcon className="h-4 w-4 text-muted-foreground dark:text-muted-foreground-night" />
+    <main className="mx-auto max-w-4xl space-y-10 px-4 py-8 sm:px-6 lg:px-8">
+      <section className="space-y-2">
+        <h2 className={PANEL_HEADING_CLASSES}>
+          <BoltIcon className={PANEL_ICON_CLASSES} />
           <span>Kill switches</span>
-        </h1>
-        <p className="text-sm text-muted-foreground dark:text-muted-foreground-night">
+        </h2>
+        <p className={PANEL_DESCRIPTION_CLASSES}>
           Control critical system functionality.
         </p>
-      </header>
 
-      {isKillSwitchesLoading ? (
-        <div className="flex justify-center py-12">
-          <Spinner />
-        </div>
-      ) : (
-        <section
-          className={cn(
-            "mt-6 rounded-2xl border border-border",
-            "bg-background shadow-sm dark:border-border-night dark:bg-background-night"
-          )}
-        >
-          {KILL_SWITCH_TYPES.map((type, index) => {
-            const {
-              title,
-              description,
-              note,
-              icon: Icon,
-            } = KILL_SWITCH_DEFINITIONS[type];
+        {isKillSwitchesLoading ? (
+          <div className="flex justify-center py-12">
+            <Spinner />
+          </div>
+        ) : (
+          <div className={PANEL_SECTION_CLASSES}>
+            {KILL_SWITCH_TYPES.map((type, index) => {
+              const {
+                title,
+                description,
+                note,
+                icon: Icon,
+              } = KILL_SWITCH_DEFINITIONS[type];
 
-            const isEnabled = enabledKillSwitches.has(type);
-            const isUpdating = updatingKillSwitch === type;
+              const isEnabled = enabledKillSwitches.has(type);
+              const isUpdating = updatingKillSwitch === type;
 
-            return (
-              <div
-                key={type}
-                className={cn(
-                  "grid gap-4 p-5 sm:grid-cols-[minmax(0,1fr)_auto] sm:items-center",
-                  index > 0 && "border-t border-border dark:border-border-night"
-                )}
-              >
-                <div className="space-y-1">
-                  <h2 className="flex items-center gap-3 text-sm font-medium text-foreground dark:text-foreground-night">
-                    <Icon className="h-4 w-4 text-foreground dark:text-foreground-night" />
-                    <span>{title}</span>
-                  </h2>
+              return (
+                <div
+                  key={type}
+                  className={cn(
+                    "grid gap-4 p-5 sm:grid-cols-[minmax(0,1fr)_auto] sm:items-center",
+                    index > 0 &&
+                      "border-t border-border dark:border-border-night"
+                  )}
+                >
+                  <div className="space-y-1">
+                    <h3 className="flex items-center gap-3 text-sm font-medium text-foreground dark:text-foreground-night">
+                      <Icon className="h-4 w-4 text-foreground dark:text-foreground-night" />
+                      <span>{title}</span>
+                    </h3>
 
-                  <p className="text-sm leading-6 text-muted-foreground dark:text-muted-foreground-night">
-                    {description}
-                  </p>
-
-                  {note && (
-                    <p className="text-xs leading-5 text-muted-foreground dark:text-muted-foreground-night">
-                      {note}
+                    <p className="text-sm leading-6 text-muted-foreground dark:text-muted-foreground-night">
+                      {description}
                     </p>
-                  )}
-                </div>
 
-                <div className="flex h-7 w-10 items-center justify-center">
-                  {isUpdating ? (
-                    <Spinner size="xs" />
-                  ) : (
-                    <SliderToggle
-                      disabled={updatingKillSwitch !== null}
-                      onClick={() => void updateKillSwitch(type, !isEnabled)}
-                      selected={isEnabled}
-                      size="xs"
-                    />
-                  )}
+                    {note && (
+                      <p className="text-xs leading-5 text-muted-foreground dark:text-muted-foreground-night">
+                        {note}
+                      </p>
+                    )}
+                  </div>
+
+                  <div className="flex h-7 w-10 items-center justify-center">
+                    {isUpdating ? (
+                      <Spinner size="xs" />
+                    ) : (
+                      <SliderToggle
+                        disabled={updatingKillSwitch !== null}
+                        onClick={() => void updateKillSwitch(type, !isEnabled)}
+                        selected={isEnabled}
+                        size="xs"
+                      />
+                    )}
+                  </div>
                 </div>
-              </div>
-            );
-          })}
-        </section>
-      )}
+              );
+            })}
+          </div>
+        )}
+      </section>
+
+      <section className="space-y-2">
+        <h2 className={PANEL_HEADING_CLASSES}>
+          <TrashIcon className={PANEL_ICON_CLASSES} />
+          <span>Sandbox Kill Requester</span>
+        </h2>
+        <p className={PANEL_DESCRIPTION_CLASSES}>
+          Mark running sandboxes for immediate reaping. The reaper or the next
+          bash invocation will destroy them and recreate fresh ones from the
+          current image.
+        </p>
+
+        {isImagesLoading ? (
+          <div className="flex justify-center py-12">
+            <Spinner />
+          </div>
+        ) : images.length === 0 ? (
+          <div className="mt-6 flex items-center gap-2 rounded-2xl border border-border bg-background p-5 text-sm text-muted-foreground dark:border-border-night dark:bg-background-night dark:text-muted-foreground-night">
+            <ExclamationCircleIcon className="h-4 w-4" />
+            <span>No registered sandbox images found.</span>
+          </div>
+        ) : (
+          <div className={PANEL_SECTION_CLASSES}>
+            {images.map(({ baseImage, version }, index) => {
+              const olderKey = sandboxKillKey(baseImage, version);
+              const allKey = sandboxKillKey(baseImage, undefined);
+              const isOlderSubmitting = submittingSandboxKill === olderKey;
+              const isAllSubmitting = submittingSandboxKill === allKey;
+
+              return (
+                <div
+                  key={olderKey}
+                  className={cn(
+                    "grid gap-4 p-5 sm:grid-cols-[minmax(0,1fr)_auto] sm:items-center",
+                    index > 0 &&
+                      "border-t border-border dark:border-border-night"
+                  )}
+                >
+                  <div className="space-y-1">
+                    <h3 className="text-sm font-medium text-foreground dark:text-foreground-night">
+                      {baseImage}
+                      <span className="text-muted-foreground dark:text-muted-foreground-night">
+                        :{version}
+                      </span>
+                    </h3>
+                    <p className="text-xs leading-5 text-muted-foreground dark:text-muted-foreground-night">
+                      Current registered version.
+                    </p>
+                  </div>
+
+                  <div className="flex items-center gap-2">
+                    <Button
+                      variant="outline"
+                      size="sm"
+                      label="Kill older versions"
+                      disabled={submittingSandboxKill !== null}
+                      isLoading={isOlderSubmitting}
+                      onClick={() =>
+                        void submitSandboxKillRequest(
+                          baseImage,
+                          version,
+                          `Request kill of all "${baseImage}" sandboxes whose version differs from "${version}"?`
+                        )
+                      }
+                    />
+                    <Button
+                      variant="warning"
+                      size="sm"
+                      label="Kill all versions"
+                      disabled={submittingSandboxKill !== null}
+                      isLoading={isAllSubmitting}
+                      onClick={() =>
+                        void submitSandboxKillRequest(
+                          baseImage,
+                          undefined,
+                          `Request kill of ALL "${baseImage}" sandboxes (every version)?`
+                        )
+                      }
+                    />
+                  </div>
+                </div>
+              );
+            })}
+          </div>
+        )}
+      </section>
     </main>
   );
 }
