+++ front/temporal/sandbox_reaper/workflows.ts
@@ -16,4 +16,4 @@ export async function sandboxReaperWorkflow(): Promise<void> {
   }
 }
 
-export { sandboxKillRequesterWorkflow } from "@app/temporal/sandbox_reaper/kill_requester/workflows";
+export { sandboxKillRequesterWorkflow } from "./kill_requester/workflows";
